:truck: Moved permission trait to a separate crate
