+++ Cargo.toml
@@ -12,6 +12,7 @@ glium = { version = "0.34", optional = true }
 piccolo = "0.3.3"
 winit = { version = "0.29", optional = true }
 strict_partial_ord_derive = { path = "./strict_partial_ord_derive"}
+permissions = { path = "./permissions" }
 
 [features]
 gui = [ "clap", "cpal", "glium", "winit" ]
