+++ permissions/src/lib.rs
@@ -1,0 +1,22 @@
+/// Trait representing any kind of permission, with varying degrees
+///
+/// [`Permission`] requires the implementor to define two levels of granted permissions:
+/// - [`all`]: As much permissions as the implementor can represent
+/// - [`none`]: Absolutely zero permission
+pub trait Permission: Eq + PartialOrd + Sized {
+    /// Construct a [`Permission`] requesting as much as the implementor can represent
+    fn all() -> Self;
+
+    /// Construct a [`Permission`] requesting absolutely zero permission
+    fn none() -> Self;
+
+    /// Check if the [`Permission`] object currently represents maximum permissions
+    fn is_all(&self) -> bool {
+        *self == Self::all()
+    }
+
+    /// Check if the [`Permission`] object currently represents empty permissions
+    fn is_none(&self) -> bool {
+        *self == Self::none()
+    }
+}
